+++ src/editor/core/draw/frame/Footer.ts
@@ -104,6 +104,7 @@ export class Footer {
   }
 
   public getHeight(): number {
+    if (this.options.footer.disabled) return 0
     const maxHeight = this.getMaxHeight()
     const rowHeight = this.getRowHeight()
     return rowHeight > maxHeight ? maxHeight : rowHeight