+++ src/main.ts
@@ -552,8 +552,6 @@ function initEditorInstance(
         const url = payload.find(p => p.name === 'url')?.value
         if (!url) return
         instance.command.executeHyperlink({
-          type: ElementType.HYPERLINK,
-          value: '',
           url,
           valueList: splitText(name).map(n => ({
             value: n,